# cmd/apps/apis/rest/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  types.ConfigItem:
    properties:
      changed_at:
        type: string
      group:
        type: string
      name:
        type: string
      value:
        type: string
    type: object
...
  v1.ConfigResponse:
    properties:
      group:
        type: string
      name:
        type: string
      value:
        type: string
    type: object
...
  v1.DeleteConfigResponse:
    properties:
      deleted:
        type: boolean
      group:
        type: string
      name:
        type: string
    type: object
...
  v1.ListConfigResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/types.ConfigItem'
        type: array
    type: object
...
  v1.SetConfigRequest:
    properties:
      value:
        type: string
    required:
    - value
    type: object
  v1.SetConfigResponse:
    properties:
      group:
        type: string
      name:
        type: string
      value:
        type: string
    type: object
...
host: localhost:7081
info:
  contact: {}
  description: NanaFS is a Reference Filing System API
  title: NanaFS API
  version: "1.0"
paths:
  /api/v1/configs/{group}:
    get:
      consumes:
      - application/json
      description: List all configuration values in a group
      parameters:
      - description: Config group
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ListConfigResponse'
      summary: List configs
      tags:
      - Configs
  /api/v1/configs/{group}/{name}:
    delete:
      consumes:
      - application/json
      description: Delete a configuration value
      parameters:
      - description: Config group
        in: path
        name: group
        required: true
        type: string
      - description: Config name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.DeleteConfigResponse'
      summary: Delete config
      tags:
      - Configs
    get:
      consumes:
      - application/json
      description: Retrieve a configuration value by group and name
      parameters:
      - description: Config group
        in: path
        name: group
        required: true
        type: string
      - description: Config name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ConfigResponse'
      summary: Get config
      tags:
      - Configs
    put:
      consumes:
      - application/json
      description: Set a configuration value
      parameters:
      - description: Config group
        in: path
        name: group
        required: true
        type: string
      - description: Config name
        in: path
        name: name
        required: true
        type: string
      - description: Set config request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.SetConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SetConfigResponse'
      summary: Set config
      tags:
      - Configs
...
swagger: "2.0"